# doc/api.yaml
openapi: 3.0.0
info:
  title : WASAphoto 
  description: |
    WASAphoto: uploading and sharing photos with your friends. API made by 1982464 .
  version : 0.0.1

########## CONTROLLARE TUTTI I TAGS PERCHÈ NON HANNO SENSO 
paths:
...
  /banned/{userid}:
    parameters:
      - $ref: "#/components/parameters/user_id"

        
    put:
      operationId: banUser
      tags: [ban]
      summary: User to be banned 
      description: ban a user
      
      responses:
        "204":
          description: User banned other user

        "400":
          description: bad request 
      
        '401':
          $ref: "#/components/responses/Unauthorized"
          
        '500': 
          $ref: "#/components/responses/Server_error"
         
         
    delete:
      operationId: unbanUser
      tags: [ban]
      summary:  User is no longer banned 
      description: no longer banned 
      
      
      responses:
        "204":
          description: User unbanned other user.

        "400":
          description: bad request 

        '401':
          $ref: "#/components/responses/Unauthorized"
          
        '500': 
          $ref: "#/components/responses/Server_error"


############           
  /following/{userid}:
    parameters:
      - $ref: "#/components/parameters/user_id"  
        
    put:
      operationId: followUser
      tags: ["following"]
      summary: "follow another user"
      description: "a single user ( that is not banned ) decide to foolow another user" 
      
      
      responses:

        "204":
          description: User started followin other user.

        "400":
          description: bad request 

        '401':
          $ref: "#/components/responses/Unauthorized"

        '403':
          $ref: '#/components/responses/Forbidden'
          
        '500': 
          $ref: "#/components/responses/Server_error"
      
        
    delete:
      operationId: unfollowUser
      tags: ["following"]
      summary: "unfollowing a peculiar user"
      description: "a user decides to stop following another user"
      
      
      responses:

        '204':
          description: user unfollowed someone correctly 

        "400":
          description: bad request
        
        '401':
          $ref: "#/components/responses/Unauthorized"
          
        '500': 
          $ref: "#/components/responses/Server_error"
...
  /users/{userid}/photos/{photoid}:  
    parameters:
      - $ref: '#/components/parameters/user_id'
      - $ref: '#/components/parameters/photo_id'
      

    get:
      operationId: getPhoto
      tags: ["photo"]
      summary: Downloads a photo
      description: user accesses its specific photo or other peculiar photo
      

      responses:
        '200':
          description: specific photo for a specific user retrieved
          content:
            image/*:
              schema: { $ref: "#/components/schemas/Raw_photo" }

        '400':
          description: invalid parameters 

        '401':
          $ref: "#/components/responses/Unauthorized"
          
          
        '500': 
          $ref: "#/components/responses/Server_error"

    delete:
      operationId: deletePhoto
      tags: [photo]
      summary: delete a pic
      description: user deletes its photo from the stream along with likes and comments 
      responses:
        "204":
          description: User deleted photo.

        "400":
          description: Invalid identifier parameter

        "401":
          $ref: '#/components/responses/Unauthorized'

        "403":
          description:  user not author of photo.

        "500":
          $ref: '#/components/responses/Server_error'
############

###### 
...
  /like/{photoid}:
    parameters:
      - $ref: '#/components/parameters/photo_id'

    
    put: 
      operationId: likePhoto
      tags: ["like"]
      summary: "put a like"
      description: through the identifier of a pic I put a like 
      
      responses:

        "204":
          description: The like has been put to the specified photo.

        "400":
          description: invalid parameter

        '401':
          $ref: "#/components/responses/Unauthorized"
          
        '500': 
          $ref: "#/components/responses/Server_error"
    
    delete:
      operationId: unlikePhoto
      tags: ["like"]
      summary: "unlike a photo" 
      description: "user delete a like from a photo"
      
  
      responses: 
        "204":
          description: The like has been removed from the specified photo.

        "400":
          description: invalid parameter

        '401':
          $ref: "#/components/responses/Unauthorized"
          
        '404': 
          $ref: "#/components/responses/Not_found"
          
        '500': 
          $ref: "#/components/responses/Server_error"
 #########     
components :
  securitySchemes:
      bearerAuth:            
        type: http
        scheme: bearer
        bearerFormat: JWT

    
      
#################################
  parameters: 
      user_id:
        name: userid 
        in: path 
        required: true
        description: identifier of the user 
        schema:
          $ref: "#/components/schemas/User_login"
      photo_id:
        name: photoid
        in: path 
        required: true
        description: identifier of the photo
        schema:
          $ref: "#/components/schemas/Photo_id"
...
  schemas:
      User_login: # USER_ID 
        description: Identifier sent by user during the login
        readOnly: true 
        type: number  
        format: uint64
        example: 24
...
      Photo_id: #PHOTOID
        description: unique identifier of a picture
        readOnly: True 
        type: number 
        format: uint64
        example: 145
...
      Raw_photo: #BINARYIMAGE
        description: Image encoding 
        type: string
        minLength: 4
        maxLength: 65535
        format: binary 
...
  responses : 
      Unauthorized:
        description: "The user is not logged in"   
      Not_found:
        description: User not found
      Server_error:
        description: Server error
      Forbidden:
        description: user in banned list
...
security:
  - bearerAuth: []